Some changes here and there.
The server isn't in charge of creating random numbers, there is an
apposite function in the protocol.c file.
I've just started to better define the state of the server by adding
some new amazing field in the associated struct.
I've started to define some functions in order to make the main less
long.